Rename gherkin_query as query
