+++ gherkin/ruby/lib/gherkin/query.rb
